Refactor to allow re-running of last test.
To be able to rerun the last test run, we need on a test run to store the
command run. When trying to run a test, if there is a test under point we run it
otherwise we run the saved test.

To share more code and to prepare for other features like having a lists of all
tests, the code is refactored so that each backend provides functions to return
a test object for a context. A context is a specific test, a file/class or a
project. The test object contains all the info needed to reconstruct the test
command. The backend also provides a function to run the test identified by the
object.