Cosmetic changes.
Remove sections ;; ... in the backends. The problems with those are that it's
never always clear what should go under one or not and at one point somebody
mess it up forget to move it to the right place and then it does not serve a
purpose anymore. Better of just grouping related function together.